requirements.txt
langchain
langchain-openai
langchain-community
duckduckgo-search
requests
python-dotenv
beautifulsoup4
streamlit